Updated to jetty-7.4.0.RC0
git-svn-id: svn+ssh://dev.eclipse.org/svnroot/rt/org.eclipse.jetty/jetty/branches/jetty-8@2996 7e9141cc-0065-0410-87d8-b60c137991c4

+++ jetty-http/src/main/java/org/eclipse/jetty/http/Parser.java
@@ -17,17 +17,18 @@ import java.io.IOException;
 
 /**
  * Abstract interface for a connection Parser for use by Jetty.
- * 
- *
  */
 public interface Parser
 {
-
     void reset(boolean returnBuffers);
 
     boolean isComplete();
 
-    long parseAvailable() throws IOException;
+    /**
+     * @return An indication of progress, typically the number of bytes filled plus the events parsed: -1 means EOF read, 0 no progress, >0 progress
+     * @throws IOException
+     */
+    int parseAvailable() throws IOException;
 
     boolean isMoreInBuffer() throws IOException;
