+++ jetty-servlets/src/main/java/org/eclipse/jetty/servlets/DoSFilter.java
@@ -733,7 +733,7 @@ public class DoSFilter implements Filter
      * 
      * @return number of requests
      */
-    public long getThrottledRequests()
+    public int getThrottledRequests()
     {
         return _throttledRequests;
     }
